emacs block-comment-mode: Fixed align-width with non centering
Previously, align-width did the same operation regardless if we are in
centering or non-centering mode. This resulted in strage behaviour for
the non-centered mode.

Added function that detects if centering or non-centering is enabled
for individual rows. This is done by looking at the diff between the
margins.

The auto detection is used together with the align-width function in
order to tailer the align-width behaviour depending on the mode of
each line