emacs: block-comment: Bugfix with width alignment
If two block comment rows looked like this:

/*   ddddd        */
/*            dd<p>  */

And point removed the two 'd':s on the lower row, the width alignment
would crash. THe reason being that it only looked at the width of the
text, not the rightmost part of the text. This was fixed.

Aligning the point when this happens was also implemented.