emacs: block-comment: Fixed width alignment in e-lisp mode
Width alignment missbehaved in elisp since both the top and bottom
enclose were identical. This bug applied to all such modes.

Problem was in jump to body start/end functions which did not take the
row type into consideration. Added helper functions that determine if
current row is body or enclose and set the appropirate prefix.